requirements.txt
Flask==2.3.2
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
pandas==2.0.1
